+++ packages/realm/src/Realm.ts
@@ -603,6 +603,8 @@ export class Realm {
   private changeListeners = new RealmListeners(this, RealmEvent.Change);
   private beforeNotifyListeners = new RealmListeners(this, RealmEvent.BeforeNotify);
   private schemaListeners = new RealmListeners(this, RealmEvent.Schema);
+  /** @internal */
+  public currentUpdateMode: UpdateMode | undefined;
 
   /**
    * Create a new {@link Realm} instance, at the default path.
@@ -858,11 +860,20 @@ export class Realm {
       throw new Error("Cannot create an object from a detached RealmObject instance");
     }
     if (!Object.values(UpdateMode).includes(mode)) {
-      throw new Error("Unsupported 'updateMode'. Only 'never', 'modified' or 'all' is supported.");
+      throw new Error(
+        `Unsupported 'updateMode'. Only '${UpdateMode.Never}', '${UpdateMode.Modified}' or '${UpdateMode.All}' is supported.`,
+      );
     }
     this.internal.verifyOpen();
     const helpers = this.classes.getHelpers(type);
-    const realmObject = RealmObject.create(this, values, mode, { helpers });
+
+    this.currentUpdateMode = mode;
+    let realmObject: RealmObject;
+    try {
+      realmObject = RealmObject.create(this, values, mode, { helpers });
+    } finally {
+      this.currentUpdateMode = undefined;
+    }
 
     return isAsymmetric(helpers.objectSchema) ? undefined : realmObject;
   }
@@ -953,7 +964,7 @@ export class Realm {
       throw new Error("You cannot query an asymmetric object.");
     }
     const table = binding.Helpers.getTable(this.internal, objectSchema.tableKey);
-    const value = properties.get(objectSchema.primaryKey).toBinding(primaryKey, undefined);
+    const value = properties.get(objectSchema.primaryKey).toBinding(primaryKey);
     try {
       const objKey = table.findPrimaryKey(value);
       if (binding.isEmptyObjKey(objKey)) {